+++ src/main/java/bbidder/Stoppers.java
@@ -6,9 +6,9 @@ import bbidder.utils.BitUtil;
 
 public final class Stoppers {
     public static Stoppers EMPTY = new Stoppers((byte) 0);
-    private final byte patt;
+    private final short patt;
 
-    public Stoppers(byte patt) {
+    public Stoppers(short patt) {
         super();
         this.patt = patt;
     }
