+++ src/main/java/bbidder/symbols/OtherMajorSymbol.java
@@ -51,7 +51,7 @@ public final class OtherMajorSymbol extends Symbol {
         if (M != null) {
             return Stream.of(new ConstSymbol(MajorSymbol.otherMajor(M)).new Context(suitTable));
         }
-        return BitUtil.stream(Constants.MAJORS & ~suitTable.getSuits())
+        return BitUtil.stream((short)(Constants.MAJORS & ~suitTable.getSuits()))
                 .mapToObj(s -> new ConstSymbol(s).new Context(suitTable.withSuitAdded(NAME, s)));
     }
